requirements.txt
torch
torchvision
numpy
pandas
Pillow
opencv-python
scipy
scikit-learn
scikit-image
matplotlib
tqdm
pytorch-msssim
albumentations
gudhi
